move MatchEndpoint func to matchutils
